Create README.md
While this project simply shows a base level of knowledge sorrounding variables, this project is not meant to showcase coding knowledge, bur rather a starting point to see how far I have come. I hope this stands as a marker to show the leaps and bounds made in my learning.